requirements.txt
pyrogram
tgcrypto
aiofiles
motor
dnspython
python-dotenv
